+++ include/ArduinoJson/JsonBufferImpl.hpp
@@ -18,21 +18,3 @@ inline ArduinoJson::JsonObject &ArduinoJson::JsonBuffer::createObject() {
   JsonObject *ptr = new (this) JsonObject(this);
   return ptr ? *ptr : JsonObject::invalid();
 }
-
-inline ArduinoJson::JsonArray &ArduinoJson::JsonBuffer::parseArray(
-    char *json, uint8_t nestingLimit) {
-  Internals::JsonParser parser(this, json, nestingLimit);
-  return parser.parseArray();
-}
-
-inline ArduinoJson::JsonObject &ArduinoJson::JsonBuffer::parseObject(
-    char *json, uint8_t nestingLimit) {
-  Internals::JsonParser parser(this, json, nestingLimit);
-  return parser.parseObject();
-}
-
-inline ArduinoJson::JsonVariant ArduinoJson::JsonBuffer::parse(
-    char *json, uint8_t nestingLimit) {
-  Internals::JsonParser parser(this, json, nestingLimit);
-  return parser.parseVariant();
-}