+++ include/ArduinoJson/StaticJsonBuffer.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "JsonBuffer.hpp"
+#include "JsonBufferBase.hpp"
 
 #if defined(__clang__)
 #pragma clang diagnostic push
@@ -21,31 +21,86 @@
 
 namespace ArduinoJson {
 
-// Implements a JsonBuffer with fixed memory allocation.
-// The template paramenter CAPACITY specifies the capacity of the buffer in
-// bytes.
-template <size_t CAPACITY>
-class StaticJsonBuffer : public JsonBuffer {
+class StaticJsonBufferBase : public JsonBufferBase<StaticJsonBufferBase> {
  public:
-  explicit StaticJsonBuffer() : _size(0) {}
+  class String {
+   public:
+    String(StaticJsonBufferBase* parent) : _parent(parent) {
+      _start = parent->_buffer + parent->_size;
+    }
+
+    void append(char c) {
+      if (_parent->canAlloc(1)) {
+        char* last = static_cast<char*>(_parent->doAlloc(1));
+        *last = c;
+      }
+    }
+
+    const char* c_str() const {
+      if (_parent->canAlloc(1)) {
+        char* last = static_cast<char*>(_parent->doAlloc(1));
+        *last = '\0';
+        return _start;
+      } else {
+        return NULL;
+      }
+    }
+
+   private:
+    StaticJsonBufferBase* _parent;
+    char* _start;
+  };
+
+  StaticJsonBufferBase(char* buffer, size_t capa)
+      : _buffer(buffer), _capacity(capa), _size(0) {}
 
   size_t capacity() const {
-    return CAPACITY;
+    return _capacity;
   }
   size_t size() const {
     return _size;
   }
 
   virtual void* alloc(size_t bytes) {
-    if (_size + bytes > CAPACITY) return NULL;
-    void* p = &_buffer[_size];
-    _size += round_size_up(bytes);
-    return p;
+    alignNextAlloc();
+    if (!canAlloc(bytes)) return NULL;
+    return doAlloc(bytes);
+  }
+
+  String startString() {
+    return String(this);
   }
 
  private:
-  uint8_t _buffer[CAPACITY];
+  void alignNextAlloc() {
+    _size = round_size_up(_size);
+  }
+
+  bool canAlloc(size_t bytes) const {
+    return _size + bytes <= _capacity;
+  }
+
+  void* doAlloc(size_t bytes) {
+    void* p = &_buffer[_size];
+    _size += bytes;
+    return p;
+  }
+
+  char* _buffer;
+  size_t _capacity;
   size_t _size;
+};
+
+// Implements a JsonBuffer with fixed memory allocation.
+// The template paramenter CAPACITY specifies the capacity of the buffer in
+// bytes.
+template <size_t CAPACITY>
+class StaticJsonBuffer : public StaticJsonBufferBase {
+ public:
+  explicit StaticJsonBuffer() : StaticJsonBufferBase(_buffer, CAPACITY) {}
+
+ private:
+  char _buffer[CAPACITY];
 };
 }